chore: Refactored frontend single screen
