+++ ui/package.json
@@ -4,6 +4,8 @@
   "private": true,
   "dependencies": {
     "@milkdown/crepe": "^7.19.0",
+    "@milkdown/kit": "^7.19.0",
+    "@milkdown/react": "^7.19.0",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.2",
