+++ ui/src/constants.js
@@ -1,0 +1,2 @@
+export const API_URL = process.env.REACT_APP_API_URL || "";
+export const WS_URL = process.env.REACT_APP_WS_URL || "";
